+++ src/core/client/stream/tabs/Profile/ChangeEmail/ChangeEmailContainer.tsx
@@ -28,7 +28,7 @@ import {
 import { ChangeEmailContainer_settings as SettingsData } from "coral-stream/__generated__/ChangeEmailContainer_settings.graphql";
 import { ChangeEmailContainer_viewer as ViewerData } from "coral-stream/__generated__/ChangeEmailContainer_viewer.graphql";
 import CLASSES from "coral-stream/classes";
-import ValidationMessageHelper from "coral-stream/common/ValidationMessageHelper";
+import FormValidationMessage from "coral-stream/common/FormValidationMessage";
 import {
   Button,
   ButtonIcon,
@@ -293,7 +293,7 @@ const changeEmailContainer: FunctionComponent<Props> = ({
                                 fullWidth
                                 id="profile-changeEmail-Email"
                               />
-                              <ValidationMessageHelper meta={meta} />
+                              <FormValidationMessage meta={meta} />
                             </>
                           )}
                         </Field>
@@ -325,7 +325,7 @@ const changeEmailContainer: FunctionComponent<Props> = ({
                                   {...input}
                                 />
                               </Localized>
-                              <ValidationMessageHelper meta={meta} fullWidth />
+                              <FormValidationMessage meta={meta} fullWidth />
                             </FormField>
                           )}
                         </Field>
